% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puller.R
\name{az_login}
\alias{az_login}
\title{Log into the Azure CLI by running the \verb{az login} command.}
\usage{
az_login(tenant, use_device_code = FALSE)
}
\arguments{
\item{tenant}{The Azure Active Directory tenant name or GUID.}

\item{use_device_code}{default FALSe. Set to TRUE if you wish to use device
code flow instead of interactive web sign-in flow.}
}
\description{
Log into the Azure CLI by running the \verb{az login} command.
}
