% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puller.R
\docType{class}
\name{AzureTokenCLI}
\alias{AzureTokenCLI}
\title{A subclass of AzureAuth::AzureToken R6 class for tokens from Azure CLI.}
\description{
A subclass of AzureAuth::AzureToken R6 class for tokens from Azure CLI.

A subclass of AzureAuth::AzureToken R6 class for tokens from Azure CLI.
}
\section{Super class}{
\code{\link[AzureAuth:AzureToken]{AzureAuth::AzureToken}} -> \code{AzureTokenCLI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AzureTokenCLI-new}{\code{AzureTokenCLI$new()}}
\item \href{#method-AzureTokenCLI-clone}{\code{AzureTokenCLI$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AzureAuth" data-topic="AzureToken" data-id="cache"><a href='../../AzureAuth/html/AzureToken.html#method-AzureToken-cache'><code>AzureAuth::AzureToken$cache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AzureAuth" data-topic="AzureToken" data-id="can_refresh"><a href='../../AzureAuth/html/AzureToken.html#method-AzureToken-can_refresh'><code>AzureAuth::AzureToken$can_refresh()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AzureAuth" data-topic="AzureToken" data-id="hash"><a href='../../AzureAuth/html/AzureToken.html#method-AzureToken-hash'><code>AzureAuth::AzureToken$hash()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AzureAuth" data-topic="AzureToken" data-id="print"><a href='../../AzureAuth/html/AzureToken.html#method-AzureToken-print'><code>AzureAuth::AzureToken$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AzureAuth" data-topic="AzureToken" data-id="refresh"><a href='../../AzureAuth/html/AzureToken.html#method-AzureToken-refresh'><code>AzureAuth::AzureToken$refresh()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AzureAuth" data-topic="AzureToken" data-id="validate"><a href='../../AzureAuth/html/AzureToken.html#method-AzureToken-validate'><code>AzureAuth::AzureToken$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzureTokenCLI-new"></a>}}
\if{latex}{\out{\hypertarget{method-AzureTokenCLI-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzureTokenCLI$new(
  resource = NULL,
  tenant = NULL,
  aad_host = "https://login.microsoftonline.com/",
  version = 1,
  use_cache = NULL,
  token_args = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{The URI of the Azure resource to which to authenticate.}

\item{\code{tenant}}{The Azure Active Directory tenant name or GUID.}

\item{\code{aad_host}}{The Azure Active Directory authority.}

\item{\code{version}}{The Azure Active Directory API version (1 or 2)}

\item{\code{use_cache}}{default TRUE. If TRUE, tokens will be cached for reuse.}

\item{\code{token_args}}{pass through args to AzureAuth token constructor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzureTokenCLI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AzureTokenCLI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzureTokenCLI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
