% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puller.R
\name{build_az_token_cmd}
\alias{build_az_token_cmd}
\title{Build the Azure CLI command to get an access token}
\usage{
build_az_token_cmd(command = "az", resource, tenant)
}
\arguments{
\item{command}{default "az"}

\item{resource}{The URI of the Azure resource to which to authenticate.}

\item{tenant}{The Azure tenant ID to authenticate to.}
}
\description{
Build the Azure CLI command to get an access token
}
