% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puller.R
\name{execute_az_cmd}
\alias{execute_az_cmd}
\title{Execute a command to the Azure CLI with arguments and handle error cases.}
\usage{
execute_az_cmd(cmd)
}
\arguments{
\item{cmd}{A list with the command in $command and a vector of arguments in
$args.}
}
\description{
Execute a command to the Azure CLI with arguments and handle error cases.
}
