% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puller.R
\name{get_az_cli_token}
\alias{get_az_cli_token}
\title{Get an access token using the Azure CLI}
\usage{
get_az_cli_token(resource, tenant, use_cache = TRUE)
}
\arguments{
\item{resource}{The URI of the Azure resource to which to authenticate.}

\item{tenant}{The Azure Active Directory tenant name or GUID.}

\item{use_cache}{default TRUE. If TRUE, tokens will be cached for reuse.}
}
\description{
Get an access token using the Azure CLI
}
\details{
Runs the \verb{az account get-access-token} command to retrieve an access
token from Azure via the Azure CLI. Presumes that you are signed in. If not,
call \code{az_login()} first.
}
