% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puller.R
\name{get_cli_token}
\alias{get_cli_token}
\title{Get an access token using the Azure CLI}
\usage{
get_cli_token(resource)
}
\arguments{
\item{resource}{The URI of the Azure resource to which to authenticate.}
}
\description{
Get an access token using the Azure CLI
}
\details{
Runs the \verb{az account get-access-token} command to retrieve an access
token from Azure via the Azure CLI. Presumes that you are signed in. If not,
call \code{az_login()} first.
}
