% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puller.R
\name{get_kusto_auth_token}
\alias{get_kusto_auth_token}
\title{Get an access token for Kusto using authorization code flow.}
\usage{
get_kusto_auth_token(cluster, tenant = "common", use_cache = TRUE)
}
\arguments{
\item{cluster}{The cluster name or URI. e.g.: "mycluster.westus"}

\item{tenant}{The Azure Active Directory tenant name or GUID.}

\item{use_cache}{default TRUE. If TRUE, tokens will be cached for reuse.}
}
\description{
Get an access token for Kusto using authorization code flow.
}
