% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puller.R
\name{handle_az_cmd_errors}
\alias{handle_az_cmd_errors}
\title{Handling for error messages returned by the Azure CLI}
\usage{
handle_az_cmd_errors(cond)
}
\arguments{
\item{cond}{the error message string}
}
\description{
Handling for error messages returned by the Azure CLI
}
