% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puller-package.R
\docType{package}
\name{puller-package}
\alias{puller}
\alias{puller-package}
\title{puller: Get data from Azure reproducibly.}
\description{
Utility functions for getting datasets from Azure databases and storage for analysis.
}
\author{
\strong{Maintainer}: Alex Kyllo \email{alex.kyllo@gmail.com}

}
\keyword{internal}
