% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puller.R
\name{query_kusto}
\alias{query_kusto}
\title{Query a Kusto server/database using an Azure CLI token by default.}
\usage{
query_kusto(server_url, database, query, token)
}
\description{
Query a Kusto server/database using an Azure CLI token by default.
}
