% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puller.R
\name{az_login}
\alias{az_login}
\title{Log into the Azure CLI by running the \verb{az login} command.}
\usage{
az_login(tenant, use_device_code = FALSE)
}
\description{
Log into the Azure CLI by running the \verb{az login} command.
}
