% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puller.R
\name{export.tbl_kusto}
\alias{export.tbl_kusto}
\title{Execute the Kusto query and export the result to Azure Storage.}
\usage{
\method{export}{tbl_kusto}(tbl, name_prefix, storage_uri, key = "impersonate", ...)
}
\description{
Execute the Kusto query and export the result to Azure Storage.
}
